requirements.txt
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.13.0
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
